Add confirmation before removal of blog

+++ src/components/Blog.js
@@ -39,14 +39,16 @@ const Blog = ({ blog, user, blogs, setBlogs }) => {
 
 		const handleClick= async (e)=>{ 
 			e.stopPropagation() 
-			const result = await blogService.deleteBlog(blog.id,user.token)
-			if(result === 204){
-				//remove this blog
-				setShowDetails(!showDetails)
-				const updatedBlogs = blogs.filter(b=> b.id !== blog.id)
-				setBlogs(updatedBlogs)
-			}
-			//add error message if not 204
+			if(window.confirm(`remove ${blog.title} by ${blog.user.name}?`)){
+				const result = await blogService.deleteBlog(blog.id,user.token)
+				if(result === 204){
+					//remove this blog
+					setShowDetails(!showDetails)
+					const updatedBlogs = blogs.filter(b=> b.id !== blog.id)
+					setBlogs(updatedBlogs)
+				}
+				//add error message if not 204
+			}//end confirm	
 		}
 		//show different things based on username
 		if(blog.user.username !== user.username){return <div></div>}
